I tried to augment the code so that fitting without PEMs is also possible. This started with a new option "no" to tmatch_type_pems in the fit_tapnet function. But it led to all sorts of nested if-statements.
I tested the new version with the old examples and it worked (I would be surprised otherwise, since I only added the PEM-free version in if-clauses), and so far it also seems to work fine without providing PEMs. That is however still rather experimental.

Also, in tapnet_packaging, I have a manual pipeline to evaluate the effect of trait-matching width for observed-trait-only networks, to check whether there is an optimum for the three networks of Tinoco (yes, is the short answer).